scripts/requirements.txt
langchain
langchain-community
langchain-anthropic
sentence-transformers
chromadb
requests
pyyaml
python-frontmatter